requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
distro==1.9.0
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
MarkupSafe==3.0.2
nltk==3.9.1
numpy==2.2.4
openai==1.68.0
pandas==2.2.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
pytz==2025.1
regex==2024.11.6
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
Werkzeug==3.1.3
flask_sqlalchemy
PyPDF2
python-dotenv
psycopg2-binary
requests
